+++ mcu/attiny-hal/src/attiny167.rs
@@ -76,3 +76,27 @@ pub mod simple_pwm {
 
     // Fixme: Implement PWM for ATtiny167.
 }
+
+pub mod spi {
+    pub use crate::periphals::spi::*;
+
+    use crate::port;
+
+    pub type Spi = avr_hal_generic::spi::Spi<
+        crate::Attiny,
+        crate::pac::SPI,
+        port::PA5,
+        port::PA4,
+        port::PA2,
+        port::PA6,
+    >;
+
+    avr_hal_generic::impl_spi! {
+        hal: crate::Attiny,
+        peripheral: crate::pac::SPI,
+        sclk: port::PA5,
+        mosi: port::PA4,
+        miso: port::PA2,
+        cs: port::PA6,
+    }
+}
