+++ mcu/attiny-hal/src/attiny85.rs
@@ -148,3 +148,9 @@ pub mod simple_pwm {
         }
     }
 }
+
+pub mod spi {
+    pub use crate::periphals::spi::*;
+
+    // Fixme: Implement SPI for ATtiny85.
+}
